requirements.txt
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Migrate==2.2.1
python-dotenv==0.10.1
sqlalchemy-utils==0.33.10
psycopg2==2.7.6.1
PyYAML==3.13